Move helpers to main file
